ApplAI App/requirements.txt
pdfplumber
PyMuPDF
python-docx
docx2txt
python-pptx
striprtf
pillow
pytesseract
openpyxl
streamlit
requests
python-dotenv
PyPDF2
openai
sentence-transformers
aiohttp
serpapi
google-search-results
pandas
faiss-cpu
numpy
pydantic 
beautifulsoup4
huggingface-hub
torch
chardet
html5lib
lxml